+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Backend unit test coverage for `integritas`.
 - Backend unit test coverage for `integritas-auth`.
 - Backend unit test coverage for `settings`.
+- Backend unit test coverage for `status`.
 
 ## [0.38.11] 2026-08-18